+++ docker-run.sh
@@ -1,0 +1,3 @@
+!#/bin/bash
+
+docker run --rm --name discord-music-bot -v $PWD:/app discord-music-bot:4
