+++ main.py.old
@@ -1,0 +1,4 @@
+import maniac
+
+if __name__ == "__main__":
+    maniac.run_bot()
